Extract selector resolution
